Added findVariantFromOptions
Created findVariantFromOptions which can be passed to the findVariant
prop of useShopifySanityProductConsolidater. It accepts an object of
options and checks the object to see if each value is valid, if it is
valid it uses it to select the product, if it is not, it uses the
default value (the first).