Updates to useShopifyData
Refactored ot utilize functions to process data rather than relying on
variables. This meant that all functions can be passed to a pipe, making
it more clear to see how the data is being processed and in what order.
It also will make it easier for processes to be added in future and at
be run at any stage easier.