Created use-shopify-data package
Migrated use-shopify-data from shopify-storefront-api-ultilities because
of its size and that its a specialized package, Included
findVariantFromOptions as it directly relates to useShopifyData.

Refactored useShopifyData to recieve a data prop that should be an array
of Shopify datasets, rather than passing explicit data. This
theoretically means that data could come from anywhere and sync with
Shopify data. Also created dataSingle which recieves a list of single
data (matching the data returned from the find prop) that will merge
with the existing data.