Major update to useShopifyData
Updated useShopifyData to be more smart. Now you can just pass in
product or productVariant data to the data prop and it will
automatically detect what data type the data is and merge it
accordingly. It detects the data type from the shopify id (which when
decoded contains the data type).