Created Backdrop component in overlay.
Created Backdrop so that a backdrop can be added by choice. This means
that without the backdrop, there will be no flex/centering styles, it
will simply be a full page invisible overlay (making it more flexible).
In ordr to add the backdrop, the backdrop component can be used which
will also include backdrop click functionality.